Helpers for computing capital ratios
These are just simple ratios and some would probably argue that
they are not needed, but I think it's important for two reasons.
Firstly, by including them it highlights their necessity of knowing
and computing them in the world of risk capital. Secondly, the
documentation for them shows people, very clearly what inputs are
needed for which ratios. There are nuances to these ratios and its
much easier when users are informed by arg names and docs. It gives
them a useful reference directly tied to their code.